use flag file to resume delete and rename
